chore: We doing setting to api
