+++ README.md
@@ -40,6 +40,9 @@ React Router DOM, React Hook Form, shadcn/ui, React Query, Vitest e mais!
         - pnpm dlx shadcn-ui@latest add dialog
         - pnpm dlx shadcn-ui@latest add card
         - pnpm dlx shadcn-ui@latest add chart
+        - pnpm i axios
+        - pnpm i zod
+        - pnpm i @tanstack/react-query
 
 # bibliotecas para serem instaladas no Visual code
     - PostCSS Language Support
