+++ pizzashop-web/src/app.tsx
@@ -1,0 +1,23 @@
+
+import { Helmet, HelmetProvider } from 'react-helmet-async'
+import { RouterProvider } from 'react-router-dom'
+import './global.css'
+import { router } from './routes'
+import { Toaster } from 'sonner'
+import { ThemeProvider } from './components/theme/theme-provider'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { queryClient } from './lib/react-query'
+
+export function App() {
+  return (
+      <HelmetProvider>
+        <ThemeProvider storageKey='pizzashop-theme' defaultTheme='dark'>
+            <Helmet titleTemplate='%s | pizza.shop'/>
+            <Toaster richColors/>
+            <QueryClientProvider client={queryClient}>
+              <RouterProvider router={router}/>
+            </QueryClientProvider>
+        </ThemeProvider>
+      </HelmetProvider>
+  )
+}
