+++ pizzashop-web/src/env.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod'
+
+const envSchema = z.object({
+    VITE_API_URL: z.string().url(),    
+})
+
+export const env = envSchema.parse(import.meta.env)
