+++ pizzashop-web/src/lib/axios.ts
@@ -1,0 +1,6 @@
+import { env } from '@/env'
+import axios from 'axios'
+
+export const api = axios.create({
+    baseURL: env.VITE_API_URL
+})
